case to unsigned before using char as index into array

+++ src/lib/unlang/xlat_expr.c
@@ -2384,7 +2384,7 @@ static ssize_t tokenize_rcode(xlat_exp_head_t *head, xlat_exp_t **out, fr_sbuff_
 	if (slen <= 0) return 0;
 
 	if (!fr_sbuff_is_terminal(&our_in, terminals)) {
-		if (!fr_dict_attr_allowed_chars[fr_sbuff_char(&our_in, '\0')]) {
+		if (!fr_dict_attr_allowed_chars[(uint8_t) fr_sbuff_char(&our_in, '\0')]) {
 			fr_strerror_const("Unexpected text after return code");
 			FR_SBUFF_ERROR_RETURN(&our_in);
 		}
